Relocate assign_req_to_token_pool triton helpers into mem_cache/allocation.py
